pyproject.toml
[project]
name = "django-scms"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asgiref==3.8.1",
    "distlib==0.3.5",
    "django==5.1",
    "django-cors-headers==4.4.0",
    "django-environ>=0.12.0",
    "django-filter==25.1",
    "djangorestframework==3.15.2",
    "djangorestframework-simplejwt==5.3.1",
    "drf-spectacular>=0.29.0",
    "environ==1.0",
    "et-xmlfile==1.1.0",
    "filelock==3.8.0",
    "openpyxl==3.1.5",
    "pillow==10.4.0",
    "platformdirs==2.5.2",
    "psycopg2-binary==2.9.10",
    "pyjwt==2.9.0",
    "python-decouple==3.8",
    "sqlparse==0.5.1",
    "typing-extensions==4.12.2",
    "tzdata==2024.1",
    "ua-parser==1.0.0",
    "ua-parser-builtins==0.18.0.post1",
    "user-agents==2.2.0",
]
